Delete account (#319)
* removed src field and fixed up delete account

* delete account updated

+++ src/main/java/com/minion/api/AccountController.java
@@ -38,8 +38,6 @@ import com.segment.analytics.messages.IdentifyMessage;
 import com.segment.analytics.messages.TrackMessage;
 import com.stripe.exception.StripeException;
 import com.stripe.model.Customer;
-import com.stripe.model.Plan;
-import com.stripe.model.Subscription;
 import com.mashape.unirest.http.HttpResponse;
 
 /**
@@ -211,7 +209,7 @@ public class AccountController {
     	Auth0Client auth = new Auth0Client();
     	String username = auth.getUsername(auth_access_token);
     	Account account = account_repo.findByUsername(username);
-    					
+    	System.err.println("Account :: " + account);
 		//remove Auth0 account
     	HttpResponse<String> response = Auth0ManagementApi.deleteUser(auth.getUserId(auth_access_token));
     	//log.info("AUTH0 Response body      :::::::::::      "+response.getBody());
@@ -220,11 +218,14 @@ public class AccountController {
     	
     	
     	//remove stripe subscription
-        this.stripeClient.cancelSubscription(account.getSubscriptionToken());
-        this.stripeClient.deleteCustomer(account.getCustomerToken());
-        
+    	if(account.getSubscriptionToken() != null && !account.getSubscriptionToken().isEmpty()){
+    		this.stripeClient.cancelSubscription(account.getSubscriptionToken());
+    	}
+    	if(account.getCustomerToken() != null && !account.getCustomerToken().isEmpty()){
+    		this.stripeClient.deleteCustomer(account.getCustomerToken());
+    	}
 		//remove account
-        account_repo.delete(account);
+        account_repo.deleteAccountAndEdges(account.getUsername());
         logger.info("update invoked");
     }
